drt: Use Bottom/Top routing layer in FlexDRWorker::initTrackCoords
When trying to find a matching layer in the other routing
direction, the search is limited to layers selected for
routing rather than all tech layers.

In addition, the code now also checks the layer is in
the actual routing direction that's expected.

Finally because that code logic was duplicated a few times,
it gets isolated in a helper function to make it easier
to follow and track what's going on.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>